myapp/requirements.txt
pandas==1.2.4
Django==3.2.7
djangorestframework==3.13.1
mssql-django
django-cors-headers